Rename the add-image-page folder to add-image
