+++ photo-sharing-application-front-end/src/app/add-image/add-image-page.component.html
